requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.52
botocore==1.34.52
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
openai==1.12.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rss-parser==2.0.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.0.7
xmltodict==0.13.0
